gdbserver: Attempt to set up the core for instruction ramming automatically
Leaving the setup to be done just before instruction ramming doesn't seem
to play nicely on P9 - the machine turns off. Doing it before we start the
gdbserver seems to work reliably.

This also has the bonus of checking that we can do instruction ramming on
the specified core before starting gdb.